Migrated tests to modules
