+++ x/bank/client/cli_test/send_test.go
@@ -1,9 +1,10 @@
-package cli
+package cli_test
 
 import (
 	"github.com/cosmos/cosmos-sdk/cli_test/helpers"
 	"github.com/cosmos/cosmos-sdk/tests"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	bankcli "github.com/cosmos/cosmos-sdk/x/bank/client/cli_test"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
@@ -21,30 +22,30 @@ func TestCLISend(t *testing.T) {
 	barAddr := f.KeyAddress(helpers.KeyBar)
 
 	startTokens := sdk.TokensFromConsensusPower(50)
-	require.Equal(t, startTokens, f.QueryBalances(fooAddr).AmountOf(helpers.Denom))
+	require.Equal(t, startTokens, bankcli.QueryBalances(f, fooAddr).AmountOf(helpers.Denom))
 
 	sendTokens := sdk.TokensFromConsensusPower(10)
 
 	// It does not allow to send in offline mode
-	success, _, stdErr := TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "-y", "--offline")
+	success, _, stdErr := bankcli.TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "-y", "--offline")
 	require.Contains(t, stdErr, "no RPC client is defined in offline mode")
 	require.False(f.T, success)
 	tests.WaitForNextNBlocksTM(1, f.Port)
 
 	// Send some tokens from one account to the other
-	TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "-y")
+	bankcli.TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "-y")
 	tests.WaitForNextNBlocksTM(1, f.Port)
 
 	// Ensure account balances match expected
-	require.Equal(t, sendTokens, f.QueryBalances(barAddr).AmountOf(helpers.Denom))
-	require.Equal(t, startTokens.Sub(sendTokens), f.QueryBalances(fooAddr).AmountOf(helpers.Denom))
+	require.Equal(t, sendTokens, bankcli.QueryBalances(f, barAddr).AmountOf(helpers.Denom))
+	require.Equal(t, startTokens.Sub(sendTokens), bankcli.QueryBalances(f, fooAddr).AmountOf(helpers.Denom))
 
 	// Test --dry-run
-	success, _, _ = TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "--dry-run")
+	success, _, _ = bankcli.TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "--dry-run")
 	require.True(t, success)
 
 	// Test --generate-only
-	success, stdout, stderr := TxSend(
+	success, stdout, stderr := bankcli.TxSend(
 		f, fooAddr.String(), barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "--generate-only=true",
 	)
 	require.Empty(t, stderr)
@@ -56,23 +57,23 @@ func TestCLISend(t *testing.T) {
 	require.Len(t, msg.GetSignatures(), 0)
 
 	// Check state didn't change
-	require.Equal(t, startTokens.Sub(sendTokens), f.QueryBalances(fooAddr).AmountOf(helpers.Denom))
+	require.Equal(t, startTokens.Sub(sendTokens), bankcli.QueryBalances(f, fooAddr).AmountOf(helpers.Denom))
 
 	// test autosequencing
-	TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "-y")
+	bankcli.TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "-y")
 	tests.WaitForNextNBlocksTM(1, f.Port)
 
 	// Ensure account balances match expected
-	require.Equal(t, sendTokens.MulRaw(2), f.QueryBalances(barAddr).AmountOf(helpers.Denom))
-	require.Equal(t, startTokens.Sub(sendTokens.MulRaw(2)), f.QueryBalances(fooAddr).AmountOf(helpers.Denom))
+	require.Equal(t, sendTokens.MulRaw(2), bankcli.QueryBalances(f, barAddr).AmountOf(helpers.Denom))
+	require.Equal(t, startTokens.Sub(sendTokens.MulRaw(2)), bankcli.QueryBalances(f, fooAddr).AmountOf(helpers.Denom))
 
 	// test memo
-	TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "--memo='testmemo'", "-y")
+	bankcli.TxSend(f, helpers.KeyFoo, barAddr, sdk.NewCoin(helpers.Denom, sendTokens), "--memo='testmemo'", "-y")
 	tests.WaitForNextNBlocksTM(1, f.Port)
 
 	// Ensure account balances match expected
-	require.Equal(t, sendTokens.MulRaw(3), f.QueryBalances(barAddr).AmountOf(helpers.Denom))
-	require.Equal(t, startTokens.Sub(sendTokens.MulRaw(3)), f.QueryBalances(fooAddr).AmountOf(helpers.Denom))
+	require.Equal(t, sendTokens.MulRaw(3), bankcli.QueryBalances(f, barAddr).AmountOf(helpers.Denom))
+	require.Equal(t, startTokens.Sub(sendTokens.MulRaw(3)), bankcli.QueryBalances(f, fooAddr).AmountOf(helpers.Denom))
 
 	f.Cleanup()
 }
